# docs/openapi.json
{
  "info": {
    "description": "Submit and manage workflows",
    "title": "REANA Workflow Controller",
    "version": "0.95.0a7"
  },
  "paths": {
...
    "/api/workflows/{workflow_id_or_name}/logs": {
      "get": {
        "description": "This resource is expecting a workflow UUID and a filename to return its outputs.",
        "operationId": "get_workflow_logs",
        "parameters": [
          {
            "description": "Required. UUID of workflow owner.",
            "in": "query",
            "name": "user",
            "required": true,
            "type": "string"
          },
          {
            "description": "Required. Workflow UUID or name.",
            "in": "path",
            "name": "workflow_id_or_name",
            "required": true,
            "type": "string"
          },
          {
            "description": "Steps of a workflow.",
            "in": "body",
            "name": "steps",
            "required": false,
            "schema": {
              "description": "List of step names to get logs for.",
              "items": {
                "description": "Step name.",
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "description": "Results page number (pagination).",
            "in": "query",
            "name": "page",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Number of results per page (pagination).",
            "in": "query",
            "name": "size",
            "required": false,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Request succeeded. Info about workflow, including the status is returned.",
            "examples": {
              "application/json": {
                "live_logs_enabled": false,
                "logs": "{'workflow_logs': string, 'job_logs': { '256b25f4-4cfb-4684-b7a8-73872ef455a2': string, '256b25f4-4cfb-4684-b7a8-73872ef455a3': string, }, 'engine_specific': object, }",
                "user": "00000000-0000-0000-0000-000000000000",
                "workflow_id": "256b25f4-4cfb-4684-b7a8-73872ef455a1",
                "workflow_name": "mytest-1"
              }
            },
            "schema": {
              "properties": {
                "live_logs_enabled": {
                  "type": "boolean"
                },
                "logs": {
                  "type": "string"
                },
                "user": {
                  "type": "string"
                },
                "workflow_id": {
                  "type": "string"
                },
                "workflow_name": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Request failed. The incoming data specification seems malformed."
          },
          "404": {
            "description": "Request failed. User does not exist.",
            "examples": {
              "application/json": {
                "message": "User 00000000-0000-0000-0000-000000000000 does not exist"
              }
            }
          },
          "500": {
            "description": "Request failed. Internal controller error.",
            "examples": {
              "application/json": {
                "message": "Internal workflow controller error."
              }
            }
          }
        },
        "summary": "Returns logs of a specific workflow from a workflow engine."
      }
    },
...
    "/api/workflows/{workflow_id_or_name}/workspace": {
      "get": {
        "description": "This resource retrieves the file list of a workspace, given its workflow UUID.",
        "operationId": "get_files",
        "parameters": [
          {
            "description": "Required. UUID of workflow owner.",
            "in": "query",
            "name": "user",
            "required": true,
            "type": "string"
          },
          {
            "description": "Required. Workflow UUID or name.",
            "in": "path",
            "name": "workflow_id_or_name",
            "required": true,
            "type": "string"
          },
          {
            "description": "File name(s) (glob) to list.",
            "in": "query",
            "name": "file_name",
            "required": false,
            "type": "string"
          },
          {
            "description": "Results page number (pagination).",
            "in": "query",
            "name": "page",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Number of results per page (pagination).",
            "in": "query",
            "name": "size",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Filter workflow workspace files by file name, size, or modification date.",
            "in": "query",
            "name": "search",
            "required": false,
            "type": "string"
          }
        ],
        "produces": [
          "multipart/form-data"
        ],
        "responses": {
          "200": {
            "description": "Requests succeeded. The list of code|input|output files has been retrieved.",
            "schema": {
              "properties": {
                "items": {
                  "items": {
                    "properties": {
                      "last-modified": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "size": {
                        "properties": {
                          "human_readable": {
                            "type": "string"
                          },
                          "raw": {
                            "type": "number"
                          }
                        },
                        "type": "object"
                      }
                    },
                    "type": "object"
                  },
                  "type": "array"
                },
                "total": {
                  "type": "integer"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Request failed. The request parameters are invalid or the filtered result set exceeds the configured display limit."
          },
          "404": {
            "description": "Request failed. Workflow does not exist.",
            "examples": {
              "application/json": {
                "message": "Workflow 256b25f4-4cfb-4684-b7a8-73872ef455a1 does not exist."
              }
            }
          },
          "500": {
            "description": "Request failed. Internal controller error.",
            "examples": {
              "application/json": {
                "message": "Internal workflow controller error."
              }
            }
          }
        },
        "summary": "Returns the workspace file list."
      },
      "post": {
        "consumes": [
          "application/octet-stream"
        ],
        "description": "This resource is expecting a workflow UUID and a file to place in the workspace.",
        "operationId": "upload_file",
        "parameters": [
          {
            "description": "Required. UUID of workflow owner.",
            "in": "query",
            "name": "user",
            "required": true,
            "type": "string"
          },
          {
            "description": "Required. Workflow UUID or name.",
            "in": "path",
            "name": "workflow_id_or_name",
            "required": true,
            "type": "string"
          },
          {
            "description": "Required. File to add to the workspace.",
            "in": "body",
            "name": "file",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Required. File name.",
            "in": "query",
            "name": "file_name",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Request succeeded. The file has been added to the workspace.",
            "examples": {
              "application/json": {
                "message": "`file_name` has been successfully uploaded."
              }
            },
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Request failed. The incoming data specification seems malformed"
          },
          "404": {
            "description": "Request failed. Workflow does not exist.",
            "examples": {
              "application/json": {
                "message": "Workflow cdcf48b1-c2f3-4693-8230-b066e088c6ac does not exist"
              }
            },
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Request failed. Internal controller error."
          }
        },
        "summary": "Adds a file to the workspace."
      }
    },
...
  },
  "swagger": "2.0"
}